feat: Implement new ticket form (main)
This commit implements a new ticket form in the dashboard. The form allows
users to create new tickets with a title and request. The form also includes
a submit button that will be used to create the ticket.

This change was made to allow users to easily create new tickets and
improve the overall user experience.